Added event config method, changed file
